Matching/requirements.txt
pandas
numpy
python-dotenv
requests
psycopg2-binary
sqlalchemy
geopy